# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Poly Twitter Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "paths": {
...
        "/users/{UserId}/posts": {
            "get": {
                "description": "Ищет посты пользователя по ID с поддержкой пагинации через параметры page и pageSize.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Поиск постов пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                        "description": "User ID (UUID)",
                        "name": "UserId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 30,
                        "minimum": 1,
                        "type": "integer",
                        "default": 15,
                        "description": "Размер страницы",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bearer \u003cjwt токен\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Посты найдены",
                        "schema": {
                            "$ref": "#/definitions/internal_features_posts_transport_http.GetPostByIdDTOResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError"
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "schema": {
                            "$ref": "#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_tryingmyb3st_PolyTweet_internal_core_domain.InternalError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_tryingmyb3st_PolyTweet_internal_core_domain.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "enum": [
                        "INVALID_REQUEST",
                        " INTERNAL_ERROR",
                        " SCHEDULE_EXISTS",
                        " SLOT_ALREADY_BOOKED",
                        " FORBIDEN",
                        " NOT_FOUND"
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
        "github_com_tryingmyb3st_PolyTweet_internal_core_domain.InternalError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "enum": [
                        "INTERNAL_ERROR"
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
...
        "internal_features_posts_transport_http.GetPostByIdDTOResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "0",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "user_id": {
                    "type": "string",
                    "x-order": "1",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "content": {
                    "type": "string",
                    "x-order": "2",
                    "example": "Hello, world!"
                },
                "parent_id": {
                    "type": "string",
                    "x-order": "3",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "reply_to": {
                    "type": "string",
                    "x-order": "4",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "image_url": {
                    "type": "string",
                    "x-order": "5",
                    "example": "https://example.com/image.jpg"
                },
                "created_at": {
                    "type": "string",
                    "x-order": "6",
                    "example": "2026-03-25T12:00:41.267Z"
                }
            }
        }
    }
}